create a 'step-utils' package and dedupe things
