+++ packages/grader/package.json
@@ -5,7 +5,8 @@
     "dependencies": {
         "@math-blocks/core": "0.0.1",
         "@math-blocks/editor-core": "0.0.1",
-        "@math-blocks/semantic": "0.0.1"
+        "@math-blocks/semantic": "0.0.1",
+        "@math-blocks/step-utils": "0.0.1"
     },
     "devDependencies": {
         "@math-blocks/testing": "0.0.1"
