+++ packages/grader/src/checks/util.ts
@@ -1,8 +1,7 @@
-import produce from "immer";
-
 import * as Semantic from "@math-blocks/semantic";
+import {Step, applySteps} from "@math-blocks/step-utils";
 
-import {Step, Context, Result} from "../types";
+import {Context, Result} from "../types";
 
 // TODO: handle negative numbers
 export const primeDecomp = (n: number): number[] => {
@@ -118,66 +117,6 @@ export const replaceNodeWithId = (
             }
         }
     }
-};
-
-const applyStep = (
-    node: Semantic.types.Node,
-    step: Step,
-): Semantic.types.Node => {
-    // Cloning is important since `Semantic.util.traverse` mutates `current`.
-    const clone = JSON.parse(JSON.stringify(node));
-
-    const path: Semantic.types.Node[] = [];
-
-    return Semantic.util.traverse(clone, {
-        enter: (node) => {
-            path.push(node);
-        },
-        exit: (node) => {
-            path.pop();
-            const parent = path[path.length - 1];
-            const oldNode = step.before;
-            const newNode = step.after;
-
-            if (
-                newNode.type === "add" &&
-                node.type !== "add" &&
-                parent?.type === "add"
-            ) {
-                // let the parent handle the application
-                return;
-            }
-            if (node.id === oldNode.id) {
-                return newNode;
-            }
-            if (newNode.type === "add" && node.type === "add") {
-                const index = node.args.findIndex(
-                    (arg) => arg.id === oldNode.id,
-                );
-                if (index !== -1) {
-                    return Semantic.builders.add([
-                        ...node.args.slice(0, index),
-                        ...newNode.args,
-                        ...node.args.slice(index + 1),
-                    ]);
-                }
-            }
-        },
-    });
-};
-
-export const applySteps = (
-    root: Semantic.types.Node,
-    steps: readonly Step[],
-): Semantic.types.Node => {
-    const nextState = produce(root, (draft) => {
-        let newState: Semantic.types.Node = draft;
-        for (const step of steps) {
-            newState = applyStep(newState, step);
-        }
-        return newState;
-    });
-    return nextState;
 };
 
 /**
