+++ packages/grader/tsconfig.json
@@ -8,6 +8,7 @@
     {"path": "../core"},
     {"path": "../editor-core"},
     {"path": "../semantic"},
+    {"path": "../step-utils"},
     {"path": "../testing"},
   ]
 }
