+++ packages/solver/src/index.ts
@@ -1,6 +1,2 @@
 export {solve} from "./solve/solve";
 export {simplify} from "./simplify/simplify";
-
-export {applyStep} from "./apply-step";
-
-export {Step} from "./types";
