+++ packages/solver/src/simplify/simplify.ts
@@ -1,6 +1,7 @@
 import * as Semantic from "@math-blocks/semantic";
+import {Step} from "@math-blocks/step-utils";
 
-import {Step, Transform} from "./types";
+import {Transform} from "./types";
 
 import {addNegToSub} from "./transforms/add-neg-to-sub";
 import {dropParens} from "./transforms/drop-parens";
