+++ packages/solver/src/simplify/transforms/collect-like-terms.ts
@@ -1,7 +1,8 @@
 import * as Semantic from "@math-blocks/semantic";
 import * as Testing from "@math-blocks/testing";
+import {Step} from "@math-blocks/step-utils";
 
-import {Step, Transform} from "../types";
+import {Transform} from "../types";
 import {simplifyMul} from "../util";
 
 import {getCoeff} from "../../solve/util";
