+++ packages/solver/src/simplify/types.ts
@@ -1,10 +1,7 @@
 import * as Semantic from "@math-blocks/semantic";
-
-import {Step} from "../types";
+import {Step} from "@math-blocks/step-utils";
 
 export type Transform = (
     node: Semantic.types.Node,
     path: Semantic.types.Node[],
 ) => Step | undefined;
-
-export {Step};
