+++ packages/step-utils/package.json
@@ -1,0 +1,9 @@
+{
+    "name": "@math-blocks/step-utils",
+    "version": "0.0.1",
+    "main": "./out/src/index.js",
+    "dependencies": {
+        "@math-blocks/core": "0.0.1",
+        "@math-blocks/semantic": "0.0.1"
+    }
+}
