+++ packages/step-utils/src/types.ts
@@ -4,5 +4,5 @@ export type Step = {
     message: string;
     before: Semantic.types.Node;
     after: Semantic.types.Node;
-    substeps: Step[];
+    substeps: readonly Step[];
 };
