+++ tsconfig-coverage.json
@@ -9,6 +9,7 @@
     {"path": "./packages/react"},
     {"path": "./packages/semantic"},
     {"path": "./packages/solver"},
+    {"path": "./packages/step-utils"},
     {"path": "./packages/testing"},
     {"path": "./packages/typesetter"},
   ]
