+++ tsconfig.json
@@ -10,6 +10,7 @@
     {"path": "./packages/react"},
     {"path": "./packages/semantic"},
     {"path": "./packages/solver"},
+    {"path": "./packages/step-utils"},
     {"path": "./packages/testing"},
     {"path": "./packages/typesetter"}
   ]
